[#119]: added map_handlers module
