+++ core/src/components/identity/src/mod.rs
@@ -1,3 +1,4 @@
 pub mod helpers;
 pub mod structs;
 pub mod enums;
+pub mod map_handlers;
